requirements.txt
# LITE - Linux Investigation & Triage Environment
# Python Dependencies

# Core Flask Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Security
Flask-Login==0.6.3
Flask-CSRF==0.9.2
Werkzeug==2.3.7

# JSON Processing
orjson==3.9.7
ijson==3.2.3

# File Handling
Watchdog==3.0.0

# Utilities
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# HTTP Requests
requests==2.31.0
urllib3==2.0.4

# Configuration
python-dotenv==1.0.1

# Logging
coloredlogs==15.0.1

# Development Tools (Optional)
Flask-DebugToolbar==0.13.1

# Testing (Optional)
pytest==7.4.2
pytest-flask==1.2.0

# Code Quality (Optional)
flake8==6.0.0
black==23.7.0

# Performance Monitoring (Optional)
Flask-APM==0.1.0

# Export Functionality
openpyxl==3.1.2
csv-writer==1.0.0

# System Information
psutil==5.9.5

# UUID Generation
uuid==1.30